Use of math.random completes phase I.
The addition of using the math.random function to determine if "red" or "green" gets to move first makes the current implementation of this game effectively functional. Initial position of pieces make require adjustment to move to the center of the closest square and resolve conflicts. Also a restart may be required if pieces are touching at the very beginning.